+++ app/src/commonMain/kotlin/it/vercruysse/lemmyapi/v1/x0/x0/datatypes/CreatePostLike.kt
@@ -5,5 +5,5 @@ import kotlinx.serialization.Serializable
 @Serializable
 internal data class CreatePostLike(
     val post_id: PostId,
-    val score: Long,
+    val score: Int,
 )
